requirements.txt
APScheduler==3.11.2
Deprecated==1.3.1
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.11.0
PyMuPDF==1.26.5
PyYAML==6.0.3
Pygments==2.19.2
SQLAlchemy==2.0.46
StrEnum==0.4.15
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
ecdsa==0.19.1
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fastapi==0.128.8
fsspec==2025.10.0
google-analytics-data>=0.18.0
google-auth>=2.0.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
limits==4.2
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==24.2
playwright==1.58.0
postgrest==2.28.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
pyiceberg==0.10.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
realtime==2.28.0
requests==2.32.5
resend==2.21.0
rich==14.3.2
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.49.3
storage3==2.28.0
strictyaml==1.7.3
supabase-auth==2.28.0
supabase-functions==2.28.0
supabase==2.28.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.1.1
yarl==1.22.0
